+++ react-client/src/actions/applicationActions.js
@@ -1,5 +1,6 @@
 import { APPLY_TO_POSITION, FETCH_ALL_APPLICATIONS } from './types';
 import { alertActions } from './alertActions';
+import axios from 'axios';
 
 // Retrieve all applicants who have applied to the job post
 export const fetchAllApplications = () => dispatch => {
@@ -15,14 +16,7 @@ export const fetchAllApplications = () => dispatch => {
 }
 
 export const applyToPosition = (application) => dispatch => {
-    const requestOptions = {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(application)
-    };
-
-    fetch('/api/application', requestOptions)
-        .then(res => res.json())
+    axios.post('/api/application', application)
         .then(
             application => {
                 dispatch({
@@ -32,32 +26,9 @@ export const applyToPosition = (application) => dispatch => {
                 //Refresh the view so that the user doesn't apply again
                 window.location.reload();
                 dispatch(alertActions.success('You have applied!'));
-            },
-            error => {
-                console.log(error);
             }
-        );
+        )
+        .catch(function(error) {
+            console.log(error);
+        });
 };
-
-
-//Currently not being used
-export const checkIfTalentApplied = (talentId, positionId) => dispatch => {
-    console.log(talentId);
-    console.log(positionId);
-    fetch(`/api/application/${talentId}/${positionId}`)
-        .then(res => res.json())
-        .then(
-            application => {
-                var payload = false;
-                if (application){
-                    payload = true;
-                }
-                // dispatch({
-                //     type: APPLIED_TO_POSITION,
-                //     payload: payload
-                // });
-                dispatch(alertActions.success('You have applied to this position.'))
-                console.log(application);
-            }
-        );
-}
