+++ react-client/src/actions/positionActions.js
@@ -1,5 +1,6 @@
 import { FETCH_ALL_POSITIONS,FETCH_EMPLOYERS_POSITIONS, FETCH_POSITION, EMPLOYER_ADD_POSITION, EMPLOYER_EDIT_POSITION } from './types';
 import history from '../helpers/history';
+import axios from 'axios';
 
 export const fetchAllPositions = () => dispatch => {
     fetch('/api/positions')
@@ -25,7 +26,6 @@ export const fetchEmployersPositions = (employerId) => dispatch => {
 };
 
 export const fetchPosition = (positionId) => dispatch => {
-    // var token = localStorage.getItem('jwtToken');
     fetch(`/api/positions/${positionId}`)
         .then(res => res.json())
         .then(position =>
@@ -37,14 +37,7 @@ export const fetchPosition = (positionId) => dispatch => {
 };
 
 export const employerAddPosition = (position) => dispatch => {
-    const requestOptions = {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(position)
-    };
-
-    fetch('/api/position', requestOptions)
-        .then(res => res.json())
+    return axios.post('/api/position', position)
         .then(
             position => {
                 dispatch({
@@ -53,17 +46,14 @@ export const employerAddPosition = (position) => dispatch => {
                 });
                 history.push('/positions');
             }
-        );
+        )
+        .catch(function (error) {
+            console.log(error);
+        });
 };
 
 export const employerEditPosition = (position, positionId) => dispatch => {
-    const requestOptions = {
-        method: 'PUT',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(position)
-    };
-    fetch(`/api/edit-position/${positionId}`, requestOptions)
-        .then(res => res.json())
+    return axios.put(`/api/edit-position/${positionId}`, position)
         .then(
             position => {
                 dispatch({
@@ -71,25 +61,22 @@ export const employerEditPosition = (position, positionId) => dispatch => {
                     payload: position
                 });
                 history.replace(`/position/${positionId}`);
-            },
-            error => {
-                console.log(error);
             }
-        );
+        )
+        .catch(function(error) {
+            console.log(error);
+
+        });
 };
 
 export const employerDeletePosition = (positionId) => dispatch => {
-    const requestOptions = {
-        method: 'DELETE'
-    };
-    fetch(`/api/delete-position/${positionId}`, requestOptions)
+    axios.delete(`/api/delete-position/${positionId}`)
         .then(
             deleted => {
-                console.log(deleted);
                 history.replace('/positions');
-            },
-            error => {
-                console.log(error);
             }
-        );
+        )
+        .catch(function(error) {
+            console.log(error);
+        });
 };
